requirements.txt
pandas
sqlalchemy
streamlit
google-genai
python-dotenv
plotly
openpyxl